+++ src/app/features/public/public.routes.ts
@@ -6,6 +6,7 @@ import {blogRoutes} from '../blog/blog.routes';
 export const publicRoutes: Routes = [
   {
     path: '',
+    pathMatch: 'full',
     data: {seo: HOME_SEO_METADATA},
     loadComponent: () => import('../../components/main/main.component').then(m => m.MainComponent),
   },